Refactor/rearrange the API wrapper into more logical directory structure
