+++ src/types/apiData.d.ts
@@ -11,7 +11,7 @@ export interface ErrorObject {
 
 /**
  *
- * Game
+ * Games
  *
  */
 
@@ -25,6 +25,40 @@ export interface ResponseGame {
   user_id: number
   name: string
   description: string | null
+  created_at: Date
+  updated_at: Date
+}
+
+/**
+ *
+ * Shopping Lists
+ *
+ */
+
+export interface ResponseShoppingList {
+  id: number
+  game_id: number
+  aggregate: boolean
+  aggregate_list_id: number
+  title: string
+  created_at: Date
+  updated_at: Date
+  list_items: ResponseShoppingListItem[]
+}
+
+/**
+ *
+ * Shopping List Items
+ *
+ */
+
+export interface ResponseShoppingListItem {
+  id: number
+  list_id: number
+  description: string
+  quantity: number
+  unit_weight: number
+  notes: string
   created_at: Date
   updated_at: Date
 }
